Finish file structure refactoring
