+++ assets/src/components/editor/commands/CmdQuote.tsx
@@ -2,7 +2,7 @@ import { ReactEditor } from 'slate-react';
 import { Transforms } from 'slate';
 import * as ContentModel from 'data/content/model';
 import { getHighestTopLevel } from 'components/editor/utils';
-import { CommandDesc, Command } from 'components/editor/toolbars/interfaces';
+import { CommandDesc, Command } from 'components/editor/commands/interfaces';
 
 const command: Command = {
   execute: (context, editor: ReactEditor) => {
