+++ assets/src/components/editor/commands/CmdTitle.tsx
@@ -1,7 +1,7 @@
 import { ReactEditor } from 'slate-react';
 import { Node, Transforms } from 'slate';
 import { getNearestBlock } from '../utils';
-import { Command, CommandDesc } from '../commands/interfaces';
+import { Command, CommandDesc } from './interfaces';
 
 const parentTextTypes = {
   p: true,
